+++ src/typeCheckPayload/index.js
@@ -5,6 +5,10 @@ export default (payload, stateTypeMap) => {
   return (typeCheckErrors.length) ? typeCheckErrors += ` \n` : typeCheckErrors;
 }
 
+const createTypeError = error => (
+` \n > ${error}`
+)
+
 const typeCheckObject = (payloadObject, objectMap, keyChain = "") => {
 
   let typeCheckObjectErrors = "";
@@ -28,23 +32,23 @@ const typeCheckObject = (payloadObject, objectMap, keyChain = "") => {
     }
 
     if ((payloadValue === null || payloadValue === undefined))
-      typeCheckObjectErrors += ` \n > The payload key "${key}" was given a value of null/undefined but values cannot be set to null/undefined with typeSafe on`
+      typeCheckObjectErrors += createTypeError(`The payload key "${key}" was given a value of null/undefined but values cannot be set to null/undefined with typeSafe on`)
     else if (initialStateType === undefined)
-      typeCheckObjectErrors += ` \n > The key "${key}" was set on the payload "${keyChain}" but was not defined on the initialState for "${keyChain}"`
+      typeCheckObjectErrors += createTypeError(`The key "${key}" was set on the payload "${keyChain}" but was not defined on the initialState for "${keyChain}"`)
     else {
       if (Array.isArray(payloadValue) && Array.isArray(initialStateType))
         typeCheckObjectErrors += typeCheckArray(payloadValue, initialStateType[0], (keyChain.length) ? keyChain : key)
       else if ((typeof initialStateType === "string" && initialStateType.startsWith("[")))
         typeCheckObjectErrors += typeCheckCanBeOneOf(key, payloadValue, initialStateType)
       else if (getValueType(payloadValueType) !== getValueType(initialStateType))
-        typeCheckObjectErrors += ` \n > The payload key "${key}" was given a value with type "${payloadValueType}" but was initially set with type "${getValueType(initialStateType)}"`
+        typeCheckObjectErrors += createTypeError(`The payload key "${key}" was given a value with type "${payloadValueType}" but was initially set with type "${getValueType(initialStateType)}"`)
       else if (payloadValueType === "object")
         typeCheckObjectErrors += typeCheckObject(payloadValue, initialStateType, ((keyChain.length) ? (keyChain + `.${key}`) :  key))
     }
   }
 
   if (keyChain.length && objectMapKeys.length) {
-    typeCheckObjectErrors += ` \n > You were missing the following keys on for the payload key "${keyChain}": ${objectMapKeys.join(", ")}`
+    typeCheckObjectErrors += createTypeError(`You were missing the following keys on for the payload key "${keyChain}": ${objectMapKeys.join(", ")}`)
   }
 
   return typeCheckObjectErrors;
@@ -65,7 +69,7 @@ const typeCheckCanBeOneOf = (payloadKey, payloadValue, initialStateType) => {
       return "";
   }
 
-  return ` \n > The key "${payloadKey}" was given "${payloadValue}" but "${payloadKey}" can only be set to one of the following: ${oneOfOptions.join(", ")}`
+  return createTypeError(`The key "${payloadKey}" was given "${payloadValue}" but "${payloadKey}" can only be set to one of the following: ${oneOfOptions.join(", ")}`)
 }
 
 
@@ -80,12 +84,12 @@ export const typeCheckArray = (payloadValue, arrayItemStructure, payloadKey) => 
     const payloadItem = payloadValue[i];
     const payloadItemType = typeof payloadItem;
     if (payloadItemType !== arrayItemType)
-      return ` \n > The array for payload key "${payloadKey}" contains a value of type "${payloadItemType}" set at position ${i}. The array for key "${payloadKey}" only supports values with type "${arrayItemType}"`
+      return createTypeError(`The array for payload key "${payloadKey}" contains a value of type "${payloadItemType}" set at position ${i}. The array for key "${payloadKey}" only supports values with type "${arrayItemType}"`)
 
     if (payloadItemType === "object") {
       const typeCheckErrors = typeCheckObject(payloadItem, arrayItemStructure, payloadKey);
       if (typeCheckErrors.length)
-        return ` \n > The array for payload key "${payloadKey}" supports objects, but the object set at position ${i} has the following errors: ${typeCheckErrors} \n`
+        return createTypeError(`The array for payload key "${payloadKey}" supports objects, but the object set at position ${i} has the following errors: ${typeCheckErrors} \n`)
     }
   }