+++ src/typeMapAndUpdateInitialStates/index.js
@@ -1,5 +1,6 @@
 import arrayOf from "../arrayOf";
 import canBeOneOf from "../canBeOneOf";
+import anyValue from "../anyValue";
 import log from "../utils/log";
 import { typeCheckArray } from "../typeCheckPayload";
 import isEqual from "react-fast-compare";
@@ -9,54 +10,15 @@ export default combinedState => {
   const typeMapsForStates = {};
 
   Object.keys(combinedState).forEach(stateKey => {
-    typeMapsForStates[stateKey] = typeMapState(stateKey, combinedState[stateKey]);
+    typeMapsForStates[stateKey] = typeMapObject(stateKey, combinedState[stateKey]);
   });
+
+  console.log(typeMapsForStates);
 
   return typeMapsForStates;
 }
 
-
-const typeMapState = (stateKey, state) => {
-
-  const stateTypeMap = {};
-
-  Object.keys(state).forEach(key => {
-
-    const value = state[key];
-    let valueType = null;
-
-    try {
-      valueType = getValueType(stateKey, key, value);
-    } catch(error) {
-      log.error(error);
-    }
-    
-    stateTypeMap[key] = valueType;
-
-    if (Array.isArray(valueType)) {
-
-      if (valueType.length !== 2) return state[key] = [];
-      
-      // valueType.length will be 2 when a default array value is passed into arrayOf()
-      // -> when it's called. It's type checked here just to ensure it passes
-      const typeCheckErrors = typeCheckArray(valueType[1], valueType[0], key);
-      if (!typeCheckErrors.length)
-        state[key] = valueType[1];
-      else {
-        state[key] = [];
-        log.error(typeCheckErrors);
-      }
-
-    } else if (typeof valueType === "string" && valueType.startsWith("[")) {
-      state[key] = JSON.parse(valueType)[0];
-    }
-  });
-
-  return stateTypeMap;
-};
-
-
-export const typeMapObject = (object, stateKey) => {
+export const typeMapObject = (stateKey, object) => {
 
   const objectMap = {};
 
@@ -93,10 +55,10 @@ export const getValueType = (stateKey, key, value) => {
   else if (isValueAnObject(value) && Object.keys(value).length === 0)
     throw new Error(`The key "${key}" for addHoney("${stateKey}") was set to an empty object. Objects must be non-empty when typeSafe is set to true when calling createHoneyPot()`)
   else if (isValueAnObject(value))
-    return typeMapObject(value, stateKey)
-  else if (typeof value === "function" && (value.toString() === arrayOf().toString() || value.toString() === canBeOneOf().toString())) {
+    return typeMapObject(stateKey, value)
+  else if (typeof value === "function" && (value.toString() === arrayOf().toString() || value.toString() === canBeOneOf().toString() || value.toString() === anyValue().toString()))
     return value(key, stateKey)
-  } else
+  else
     return typeof value
 }
 
@@ -105,9 +67,61 @@ export const isValueNull = value => (
 )
 
 export const isValueAnObject = value => (
-  typeof value === "object"
+  typeof value === "object" && !Array.isArray(value)
 )
 
 export const isValueAnArray = value => (
   Array.isArray(value)
 )
+
+
+
+
+
+// 
+// const typeMapState = (stateKey, state) => {
+// 
+//   const stateTypeMap = {};
+// 
+//   Object.keys(state).forEach(key => {
+// 
+//     const value = state[key];
+//     let valueType = null;
+// 
+//     try {
+//       valueType = getValueType(stateKey, key, value);
+//     } catch(error) {
+//       log.error(error);
+//     }
+// 
+//     console.log(valueType);
+//     
+//     stateTypeMap[key] = valueType;
+// 
+//     if (typeof valueType === "object") {
+// 
+//       const type = valueType.__honeyType;
+//       console.log(type);
+//     }
+// 
+// //     if (Array.isArray(valueType)) {
+// // 
+// //       if (valueType.length !== 2) return state[key] = [];
+// //       
+// //       // valueType.length will be 2 when a default array value is passed into arrayOf()
+// //       // -> when it's called. It's type checked here just to ensure it passes
+// //       const typeCheckErrors = typeCheckArray(valueType[1], valueType[0], key);
+// //       if (!typeCheckErrors.length)
+// //         state[key] = valueType[1];
+// //       else {
+// //         state[key] = [];
+// //         log.error(typeCheckErrors);
+// //       }
+// // 
+// //     } else if (typeof valueType === "string" && valueType.startsWith("[")) {
+// //       state[key] = JSON.parse(valueType)[0];
+// //     }
+//   });
+// 
+//   return stateTypeMap;
+// };